Introduce PyClass trait and PyClassShell
