+++ src/ffi/mod.rs
@@ -1,6 +1,6 @@
 #![allow(non_camel_case_types, non_snake_case, non_upper_case_globals)]
 #![cfg_attr(Py_LIMITED_API, allow(unused_imports))]
-#![cfg_attr(feature="cargo-clippy", allow(clippy::inline_always))]
+#![cfg_attr(feature = "cargo-clippy", allow(clippy::inline_always))]
 
 pub use self::bltinmodule::*;
 pub use self::boolobject::*;
