+++ src/lib.rs
@@ -124,11 +124,12 @@ pub use crate::conversion::{
 };
 pub use crate::err::{PyDowncastError, PyErr, PyErrArguments, PyErrValue, PyResult};
 pub use crate::gil::{init_once, GILGuard, GILPool};
-pub use crate::instance::{AsPyRef, ManagedPyRef, Py, PyNativeType, PyRef, PyRefMut};
+pub use crate::instance::{ManagedPyRef, Py, PyNativeType};
 pub use crate::object::PyObject;
 pub use crate::objectprotocol::ObjectProtocol;
+pub use crate::pyclass::{PyClass, PyClassAlloc, PyClassShell};
 pub use crate::python::{prepare_freethreaded_python, Python};
-pub use crate::type_object::{PyObjectAlloc, PyRawObject, PyTypeInfo};
+pub use crate::type_object::{PyConcreteObject, PyTypeInfo};
 
 // Re-exported for wrap_function
 #[doc(hidden)]
@@ -146,9 +147,6 @@ pub use libc;
 #[doc(hidden)]
 pub use unindent;
 
-/// Raw ffi declarations for the c interface of python
-pub mod ffi;
-
 pub mod buffer;
 #[doc(hidden)]
 pub mod callback;
@@ -158,6 +156,8 @@ mod conversion;
 pub mod derive_utils;
 mod err;
 pub mod exceptions;
+/// Raw ffi declarations for the c interface of python
+pub mod ffi;
 pub mod freelist;
 mod gil;
 mod instance;
@@ -166,6 +166,7 @@ pub mod marshal;
 mod object;
 mod objectprotocol;
 pub mod prelude;
+pub mod pyclass;
 mod python;
 pub mod type_object;
 pub mod types;
