+++ src/types/any.rs
@@ -1,7 +1,8 @@
 use crate::conversion::AsPyPointer;
+use crate::conversion::PyTryFrom;
 use crate::err::PyDowncastError;
 use crate::internal_tricks::Unsendable;
-use crate::{ffi, PyObject, PyRef, PyRefMut, PyTryFrom, PyTypeInfo};
+use crate::{ffi, PyObject};
 
 /// Represents a python's [Any](https://docs.python.org/3/library/typing.html#typing.Any) type.
 /// We can convert all python objects as `PyAny`.
@@ -39,23 +40,5 @@ impl PyAny {
         T: for<'gil> PyTryFrom<'gil>,
     {
         T::try_from_mut(self)
-    }
-}
-
-impl<'a, T> From<PyRef<'a, T>> for &'a PyAny
-where
-    T: PyTypeInfo,
-{
-    fn from(pref: PyRef<'a, T>) -> &'a PyAny {
-        unsafe { &*(pref.as_ptr() as *const PyAny) }
-    }
-}
-
-impl<'a, T> From<PyRefMut<'a, T>> for &'a PyAny
-where
-    T: PyTypeInfo,
-{
-    fn from(pref: PyRefMut<'a, T>) -> &'a PyAny {
-        unsafe { &*(pref.as_ptr() as *const PyAny) }
     }
 }
